Move util crate unit tests into a new unit_tests crate
