+++ tests/unit/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "unit_tests"
+version = "0.0.1"
+authors = ["The Servo Project Developers"]
+
+[lib]
+name = "unit_tests"
+path = "lib.rs"
+doctest = false
+
+[dependencies.util]
+path = "../../components/util"
+
+[dependencies.geom]
+git = "https://github.com/servo/rust-geom"
