+++ tests/unit/lib.rs
@@ -1,0 +1,4 @@
+extern crate geom;
+extern crate util;
+
+#[cfg(test)] #[path="util/mod.rs"] mod util_tests;
